refactor: 페이지네이션 컴포넌트 모듈화 및 페이지 계산 로직 수정
